+++ lib/output/http_client.go
@@ -55,13 +55,10 @@ message has multiple parts the request will be sent according to
 
 ### Propagating Responses
 
-EXPERIMENTAL: It's possible to propagate the response from each HTTP request
-back to the input source by setting ` + "`propagate_response` to `true`" + `.
-Only inputs that support [synchronous responses](../sync_responses.md) are able
-to make use of these propagated responses.
-
-This feature is considered experimental and is therefore subject to change
-outside of major version releases.`,
+It's possible to propagate the response from each HTTP request back to the input
+source by setting ` + "`propagate_response` to `true`" + `. Only inputs that
+support [synchronous responses](../sync_responses.md) are able to make use of
+these propagated responses.`,
 	}
 }