+++ lib/output/sync_response.go
@@ -35,9 +35,6 @@ func init() {
 			return NewWriter(TypeSyncResponse, roundtrip.Writer{}, logger, stats)
 		},
 		description: `
-EXPERIMENTAL: This component is considered experimental and is therefore subject
-to change outside of major version releases.
-
 Returns the final message payload back to the input origin of the message, where
 it is dealt with according to that specific input type.